Replace most instances of type punning with an actually legal method
This should not make any difference in practice. MSVC as the only configured compiler doesn't enforce strict aliasing in the first place, but it's better to be more standards compliant.
The new pun_cast<>() helper template pretty much results in the same assembly on a sane, optimizing compiler.

Affected is all type punning used for window messages, config value stuffing, as well as the classic fast invSqrt() in Vectors.h.
Some questionable (but most likely actually fine) casts in ImGui implementation/backend code have also been adjusted.

That is a lot of puns.