Fix incorrect dashboard origin position when Desktop+ isn't the active dashboard tab
When possible we're now using the GamepadUI overlay as the reference for dashboard positioning.
This also makes the recent dashboard position workaround unnecessary in those cases and more stable against further changes.

For the legacy dashboard overlay the issue isn't fully fixed. It's mitigated for most cases, but stuck between either fixing the rotation by hand or not following the dashboard tilt with an always-upright overlay rotation (something done in the past).
The former was chosen for now.