Add alternative cursor rendering setting
Some users appear to run into issues with the hardware cursor returned by Desktop Duplication lately. The source of the issue doesn't seem to be within Desktop+, but this setting aims to offer a workaround.

The alternative render method/SoftwareCursorGrabber is a drop-in that has the same compatibility as the main method, minus animated cursors (do not seem to be properly supported in the Win32 API).

Some wording choices around cursors for the German translation have also been adjusted while at it.